pyproject.toml
[project]
name = "cc-transaction-logger-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.17"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "gspread>=6.2.1",
    "gspread-formatting>=1.2.1",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
